+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.1] - 2025-03-15
+### Fixed
+- Return type of consumer function of `peek`, `andThen`, `onSuccess`, `onFailure`, `andFinally` in the Try class
+- Return type of `run` method in the Option class
+
 ## [1.1.0] - 2025-03-02
 ### Fixed
 - run() will return Option instance being run